3.Chef/backend/requirements.txt
flask_session==0.6.0
identity==0.5.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Dance==7.0.1
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.2.1
URLObject==2.4.3
Werkzeug==3.0.1